requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
itsdangerous==2.1.2
pymongo==4.6.1
python-dotenv==1.0.0
tqdm==4.66.1
